Add files via upload
Some difficulties I had was figuring out how to convert my real life logic into block code. This is because I had to find the right commands and combine them in a way that would function correctly. Another problem was searching for the block code, because there were many blocks in the categories. Finding the one I wanted was hard.